avoid schema conflicts by dumping enum types with fully qualified name in PostgreSQL
If there are multiple enum types with the same name, but in different schemas, the `:enum_type` option dumped as part of `db:schema:dump` will be ambiguous, potentially resulting in the wrong type being used when loading the schema.

Now, if the enum type is not already fully qualified, and multiple schemas are being dumped, the type includes the schema name.