Support but deprecate `skip_before_action :verify_authenticity_token`
In #56350, `verify_authenticity_token` was renamed to
`verify_request_for_forgery_protection`, to more accurately reflect its
purpose, now that  an authenticity token is not necessarily verified.

However, because many people rely on `skip_before_action
:verify_authenticity_token`, to opt out of forgery protection, we need
to keep this working and deprecate it.  We simply mark it as having ran,
as part of `protect_from_forgery`, and when verifying the request, we
check if the method ran. If it didn't, it's because it was skipped on
its own and not via `skip_forgery_protection`, so we can emit the
deprecation warning.