Ensure `Rails::TestUnitRailtie` is defined and `config.generators.system_tests` is set when generating a system test.
For integration tests we added check for `Rails::TestUnitRailtie`.

When `rails new --skip-test` is called, we put `config.generators.system_tests = nil` in `config/application.rb`:
https://github.com/rails/rails/blob/61b5c7e2953fd0ccfab151b233ce69283c567a34/railties/lib/rails/generators/rails/app/templates/config/application.rb.tt#L36-L39

If the user then tries to run `bin/rails g system_test Whatever` or `bin/rails g integration_test Whatever` it will do nothing.

This change will now raise an exception in that case to give the user some feedback.

---

Originally I was going to leave out the const check because it creates a dependency on `Rails::TestUnit`, but I checked and `rspec-rails` is not using this generator -- they provide their own `bin/rails g rspec:integration` and `bin/rails g rspec:system`.

So the behavior is the same if you are using `rspec-rails` or not, if `rails/test_unit/railtie` is not loaded.

I couldn't find any other uses of the generator in the wilds, and I'm unaware of other test frameworks that could be affected by this change.