Yield key to `ActionController::Parameters#fetch`
Prior to this commit, the `ActionController::Parameters#fetch` method
did not yield the key to the block when fetching an unknown key. While
the `ActionController::Parameters` interface is Hash-like, it **does
not** inherit from `Hash` directly.

The `Hash#fetch` method yields the key to the block, so this commit
changes the `ActionController::Parameters#fetch` behavior to do the
same.

[Hash#fetch]: https://docs.ruby-lang.org/en/master/Hash.html#method-i-fetch