Change default from IPv4 to IPv6
IPv4 addresses like 127.0.0.1 and 0.0.0.0 are easy and familiar, but they're increasingly more expensive with provider such as AWS (https://aws.amazon.com/blogs/aws/new-aws-public-ipv4-address-charge-public-ip-insights/). IPv6 addresses are much more plentiful, and therefore cheaper (and represent the future).

This change:

- `0.0.0.0` => `::` (production)
- `127.0.0.1` => `::1` (development)


When Puma binds to these IPv6 addresses, it also works with the old IPv4 equivalents, which means someone can still type in 127.0.0.1:9292 into their browser and have it work, but the opposite is not true.

I have a corresponding issue to update the defaults in Puma to IPv6 as well. https://github.com/puma/puma/issues/3812